# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp Gateway API
  version: "1.0.0"
  description: |
    Self-hosted WhatsApp HTTP API built on tulir/whatsmeow.

    **Auth:** every `/api/*` endpoint except `/api/login` requires `Authorization: Bearer <key>` (or `X-Api-Key: <key>`). Get a key via `POST /api/login`.

    **Session ID** is the name you created (e.g. `default`). Pass it either in the path
    (`/api/{session}/...`) or as the `session` field in JSON bodies of send-message calls.

    **chat_id** accepts either a bare E.164 phone number (`6281234567890`) or a full JID
    (`6281234567890@s.whatsapp.net` for personal, `1234567890-1700000000@g.us` for groups).
servers:
  - url: http://localhost:3000
    description: Local
security:
  - BearerAuth: []
  - ApiKey: []
...
paths:
...
  /api/{session}/groups:
    parameters:
      - $ref: "#/components/parameters/SessionInPath"
    get:
      tags: [Groups]
      summary: List joined groups
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    jid:     { type: string, example: "1234567890-1700000000@g.us" }
                    name:    { type: string }
                    topic:   { type: string }
                    members: { type: integer }
    post:
      tags: [Groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, participants]
              properties:
                name:         { type: string, example: "Project Falcon" }
                participants: { type: array, items: { type: string }, example: ["6281234567890", "60123456789"] }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupInfo" }
...
  /api/{session}/groups/{gid}:
    parameters:
      - $ref: "#/components/parameters/SessionInPath"
      - $ref: "#/components/parameters/GroupID"
    get:
      tags: [Groups]
      summary: Get full group info (members, settings)
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupInfo" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
...
    SessionInPath:
      in: path
      name: session
      required: true
      schema: { type: string }
      example: default
    GroupID:
      in: path
      name: gid
      required: true
      description: "Group JID (e.g. 1234567890-1700000000@g.us) or bare group ID."
      schema: { type: string }
      example: "1234567890-1700000000@g.us"
  schemas:
...
    GroupInfo:
      type: object
      properties:
        jid:        { type: string, example: "1234567890-1700000000@g.us" }
        name:       { type: string }
        topic:      { type: string }
        locked:     { type: boolean }
        announce:   { type: boolean }
        owner:      { type: string }
        created_at: { type: integer }
        participants:
          type: array
          items:
            type: object
            properties:
              jid:          { type: string }
              is_admin:     { type: boolean }
              is_superadmin: { type: boolean }
...